+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Line wrap the file at 100 chars.                                              Th
 - Add translations for the current location displayed on the main screen in the GUI.
 - Allow a subset of NDP (Router solicitation, router advertisement and redirects) in the firewall.
 - Allow setting proxy mode from UI.
+- Automatically include frontend logs in problem report when ran from CLI.
 
 #### Linux
 - Add standard window decorations to the application window.